Divide solver files
